move letters download connection
