+++ lib/evss/letters/service.rb
@@ -22,42 +22,6 @@ module EVSS
         end
       end
 
-      def download_letter(type, options = nil)
-        with_monitoring do
-          headers = headers_for_user(@user)
-          response = make_download_request(headers, options, type)
-
-          case response.status.to_i
-          when 200
-            response.body
-          when 404
-            raise Common::Exceptions::RecordNotFound, type
-          else
-            log_message_to_sentry(
-              'EVSS letter generation failed', :error, extra_context: {
-                url: config.base_path, body: response.body
-              }
-            )
-            raise_backend_exception('EVSS502', 'Letters')
-          end
-        end
-      end
-
-      def make_download_request(headers, options, type)
-        if options.blank?
-          response = download_conn.get type do |request|
-            request.headers.update(headers)
-          end
-        else
-          response = download_conn.post do |request|
-            request.url "#{type}/generate"
-            request.headers['Content-Type'] = 'application/json'
-            request.body = options
-          end
-        end
-        response
-      end
-
       private
 
       def handle_error(error)
@@ -68,16 +32,6 @@ module EVSS
           raise EVSS::Letters::ServiceException, error.body
         else
           super(error)
-        end
-      end
-
-      # TODO(AJD): move to own service
-      def download_conn
-        @download_conn ||= Faraday.new(config.base_path, ssl: config.ssl_options) do |faraday|
-          faraday.options.timeout = EVSS::Letters::Configuration::DEFAULT_TIMEOUT
-          faraday.use :breakers
-          faraday.use EVSS::ErrorMiddleware
-          faraday.adapter Faraday.default_adapter
         end
       end
     end
